requirements.txt
faker==24.0.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
xgboost==2.0.3
lightgbm==4.3.0
prophet==1.1.5
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.9
slowapi==0.1.9
python-dotenv==1.0.1
rich==13.7.1
click==8.1.7
shap==0.44.1
matplotlib==3.8.3
seaborn==0.13.2
scipy==1.12.0
pytest==8.0.2
pytest-asyncio==0.23.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.0
requests==2.32.3
prometheus-client==0.20.0
apscheduler==3.10.4
cryptography==42.0.5
reportlab==4.2.0
